v2/backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database & Migrations
SQLAlchemy==2.0.23
alembic==1.13.1

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
werkzeug==3.0.1
argon2-cffi==23.1.0

# Email & SMS
twilio==8.11.0
python-dotenv==1.0.0

# PDF Generation
reportlab==4.0.7

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Utilities
python-slugify==8.0.1

# Development
pytest==7.4.3
pytest-cov==4.1.0
